+++ .helix/languages.toml
@@ -1,18 +1,9 @@
-# .helix/languages.toml
-
-[language-server.deno-lsp]
-command = "deno"
-args = ["lsp"]
-
-[language-server.deno-lsp.config.deno]
-enable = true
-lint = true
-unstable = true
-suggest.imports.hosts = { "https://deno.land" = true }
-
 [[language]]
 name = "typescript"
-roots = ["deno.json", "deno.jsonc"]
-language-servers = [ "deno-lsp" ]
+language-servers = ["typescript-language-server"]
 auto-format = true
 
+[[language]]
+name = "tsx"
+language-servers = ["typescript-language-server"]
+auto-format = true
