+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+  "useTabs": false,
+  "printWidth": 80,
+  "tabWidth": 2,
+  "semi": true,
+  "singleQuote": false,
+  "proseWrap": "preserve"
+}
