--- deno.json
@@ -1,59 +1,0 @@
-{
-  "export": "out",
-  "nodeModulesDir": "auto",
-  "unstable": [
-    "unsafe-proto"
-  ],
-  "imports": {
-    "next": "npm:next@16.2.2",
-    "next/": "npm:next@16.2.2/"
-  },
-  "compilerOptions": {
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "strict": true,
-    "jsx": "preserve"
-  },
-  "lint": {
-    "plugins": [
-      "jsr:@intility/no-inline-styles",
-      "./custom_lint_rules/no-inline-jsx.ts",
-      "./custom_lint_rules/single_component_plugin.ts",
-      "./custom_lint_rules/no-complex-inline-jsx.ts"
-    ],
-    "exclude": [
-      "supabase/"
-    ],
-    "rules": {
-      "tags": [
-        "recommended",
-        "react"
-      ],
-      "include": [
-        "camelcase",
-        "no-implicit-declare-namespace-export",
-        "no-type-assertion"
-      ],
-      "exclude": [
-        "no-process-global"
-      ]
-    }
-  },
-  "fmt": {
-    "useTabs": false,
-    "lineWidth": 80,
-    "indentWidth": 2,
-    "semiColons": true,
-    "singleQuote": false,
-    "proseWrap": "preserve"
-  },
-  "tasks": {
-    "dev": "deno run -A --unstable-detect-cjs npm:next@16.2.2 dev",
-    "build": "deno run -A --unstable-detect-cjs npm:next@16.2.2 build",
-    "build:pages": "deno run -A --unstable-detect-cjs npm:next@16.2.2 build",
-    "lint": "deno lint --fix && deno fmt --check"
-  }
-}
